moved files
 - Version and CertHandlingStrategy into auth.properties.property
 - AdvancedAuthenticatorActivity and AuthenticatorActivity and CertHolder into ui.auth
