+++ moVirt/src/main/res/layout/activity_advanced_authenticator.xml
@@ -7,7 +7,7 @@
                 android:paddingLeft="@dimen/activity_horizontal_margin"
                 android:paddingRight="@dimen/activity_horizontal_margin"
                 android:paddingTop="@dimen/activity_vertical_margin"
-                tools:context=".ui.AdvancedAuthenticatorActivity">
+                tools:context=".ui.auth.AdvancedAuthenticatorActivity">
 
     <ProgressBar
         android:id="@+id/progress"
