+++ moVirt/src/main/res/layout/activity_authenticator.xml
@@ -7,7 +7,7 @@
                 android:paddingLeft="@dimen/activity_horizontal_margin"
                 android:paddingRight="@dimen/activity_horizontal_margin"
                 android:paddingTop="@dimen/activity_vertical_margin"
-                tools:context=".ui.AuthenticatorActivity">
+                tools:context=".ui.auth.AuthenticatorActivity">
 
     <ProgressBar
         android:id="@+id/authProgress"
